Separate code of Scorer interface and scorer implementations + add scorerManager
